runner-flux/requirements.txt
fastapi==0.111.1
loguru==0.7.2
httpx==0.27.2
redis[async]==5.0.7
uvicorn==0.30.6
Pillow==10.4.0
